refactor(gallery): split file service to different services
split file service to different services

MODULE-2-PART-3-express
