+++ backend/src/helper/file.logger.ts
@@ -1,27 +1,27 @@
 import { writeFile } from 'node:fs/promises';
 import { Logger } from './abstract.logger.js';
-import { FileService } from '../services/file.service.js';
+import { FileSystemService } from '../services/file.system.service.js';
 
-const fileService = new FileService();
+const fsService = new FileSystemService();
 
 export class FileLogger extends Logger {
   async writeLog(level: string, message: string) {
     const content = this.formatLogContent(level, message);
 
-    if (!fileService.directoryExists(this.directory)) {
-      await fileService.createDirectory('logs');
+    if (!fsService.directoryExists(this.directory)) {
+      await fsService.createDirectory('logs');
       this.directory = 'logs';
       return;
     }
 
-    if (!fileService.fileExists(this.filePath)) {
-      await fileService.createLogFile(this.filePath, content);
+    if (!fsService.fileExists(this.filePath)) {
+      await fsService.createFile(this.filePath, content);
     }
 
     if (this.logTimeExpired()) {
       const date = new Date();
       const filePath = `./${this.directory}/logs_${date}.txt`;
-      await fileService.createLogFile(filePath, content);
+      await fsService.createFile(filePath, content);
 
       this.fileCreated = date;
       this.fileName = `logs_${date}.txt`;
