+++ backend/src/services/file.service.ts
@@ -1,7 +1,6 @@
 import fs from 'fs';
 import { log } from '../helper/logger.js';
 import { Request, Response } from 'express';
-import { mkdir, writeFile } from 'node:fs/promises';
 
 export class FileService {
   async sendFile(req: Request, res: Response, path: string, contentType: string): Promise<void> {
@@ -17,21 +16,5 @@ export class FileService {
       res.writeHead(200);
       res.end(data);
     });
-  }
-
-  fileExists(filePath: string): boolean {
-    return fs.existsSync(filePath);
-  }
-
-  directoryExists(directory: string) {
-    return fs.existsSync(directory);
-  }
-
-  async createLogFile(file: string, content: string) {
-    await writeFile(file, content, {flag: 'a+'});
-  }
-
-  async createDirectory(dirName: string) {
-    await mkdir(dirName, {recursive: true});
   }
 }
